Fix #84: Handle double semicolons in service options
When encountering a double semicolon in a service option, the parser now correctly creates an EmptyStatement struct for the second semicolon and adds it to the ServiceBody slice. This prevents a nil pointer dereference when the Service.Accept method is called.

The fix was implemented by adding a line to create an EmptyStatement struct in the default case of the switch statement in the parseServiceBody function.